Move autocheckin from plugin to service
