+++ src/pretix/base/services/checkin.py
@@ -1,11 +1,13 @@
 from django.db import transaction
 from django.db.models import Prefetch
+from django.dispatch import receiver
 from django.utils.timezone import now
 from django.utils.translation import ugettext as _
 
 from pretix.base.models import (
     Checkin, CheckinList, Order, OrderPosition, Question, QuestionOption,
 )
+from pretix.base.signals import order_placed
 
 
 class CheckInError(Exception):
@@ -155,3 +157,16 @@ def perform_checkin(op: OrderPosition, clist: CheckinList, given_answers: dict, 
             'datetime': dt,
             'list': clist.pk
         }, user=user, auth=auth)
+
+
+@receiver(order_placed, dispatch_uid="autocheckin_order_placed")
+def order_placed(sender, **kwargs):
+    order = kwargs['order']
+    event = sender
+
+    cls = list(event.checkin_lists.filter(auto_checkin_sales_channels__contains=order.sales_channel).prefetch_related(
+        'limit_products'))
+    for op in order.positions.all():
+        for cl in cls:
+            if cl.all_products or op.item_id in {i.pk for i in cl.limit_products.all()}:
+                Checkin.objects.create(position=op, list=cl, auto_checked_in=True)
