+++ src/tests/plugins/test_checkinlist.py
@@ -109,8 +109,8 @@ def test_csv_order_by_name_parts(event):  # noqa
 "Attendee name: First name","Attendee name: Middle name","Attendee name: Family name","Product","Price",
 "Checked in","Automatically checked in","Secret","E-mail","Company","Voucher code","Order date","Requires special
  attention","Comment"
-"FOO","Mr Peter A Jones","Mr","Peter","A","Jones","Ticket","23.00","","","hutjztuxhkbtwnesv2suqv26k6ttytxx",
+"FOO","Mr Peter A Jones","Mr","Peter","A","Jones","Ticket","23.00","","No","hutjztuxhkbtwnesv2suqv26k6ttytxx",
 "dummy@dummy.test","","","2019-02-22","No",""
-"FOO","Mrs Andrea J Zulu","Mrs","Andrea","J","Zulu","Ticket","13.00","","","ggsngqtnmhx74jswjngw3fk8pfwz2a7k",
+"FOO","Mrs Andrea J Zulu","Mrs","Andrea","J","Zulu","Ticket","13.00","","No","ggsngqtnmhx74jswjngw3fk8pfwz2a7k",
 "dummy@dummy.test","","","2019-02-22","No",""
 """)
